requirements.txt
# ==========================
# 🧠 CORE BACKEND
# ==========================
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
PyJWT==2.10.1
pydantic[email]
pydantic_settings
google-auth

# ==========================
# 💾 DATABASE & CLOUD
# ==========================
pymongo
boto3

# ==========================
# 🎙️ AUDIO PROCESSING
# ==========================
librosa==0.10.0.post2
numpy==1.25.2
scipy==1.11.1
pydub
torch
silero_vad

# ==========================
# 🧠 AI / LLM / SPEECH APIs
# ==========================
openai
assemblyai
torchcodec


# ==========================
# 📁 FILE HANDLING
# ==========================
aiofiles
python-multipart

# ==========================
# 🎬 VIDEO / VISION
# ==========================
moviepy==1.0.3
opencv-python
